refactor: move create mode input handling
