+++ src/app/create_device.rs
@@ -1,11 +1,124 @@
-use super::{App, AppState, Mode, Panel};
+use super::{state, App, AppState, Mode, Panel};
 use crate::constants::performance::DETAIL_UPDATE_DEBOUNCE;
 use crate::managers::common::{DeviceConfig, DeviceManager};
 use crate::models::error::format_user_error;
 use anyhow::Result;
+use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
 use std::sync::Arc;
 
 impl App {
+    pub(super) async fn enter_create_device_mode(&mut self) {
+        let active_panel = {
+            let mut state = self.state.lock().await;
+            let active_panel = state.active_panel;
+            state.mode = Mode::CreateDevice;
+            state.create_device_form = match active_panel {
+                Panel::Android => state::CreateDeviceForm::for_android(),
+                Panel::Ios => state::CreateDeviceForm::for_ios(),
+            };
+            state.create_device_form.is_loading_cache = true;
+            active_panel
+        };
+
+        let cache_available = {
+            let state = self.state.lock().await;
+            state.is_cache_available(active_panel).await
+        };
+
+        if cache_available {
+            let mut state = self.state.lock().await;
+            state.populate_form_from_cache(active_panel).await;
+            state.create_device_form.is_loading_cache = false;
+            return;
+        }
+
+        let state_clone = Arc::clone(&self.state);
+        let android_manager = self.android_manager.clone();
+        let ios_manager = self.ios_manager.clone();
+
+        tokio::spawn(async move {
+            match active_panel {
+                Panel::Android => {
+                    if let Ok(targets) = android_manager.list_available_targets().await {
+                        if let Ok(devices) =
+                            android_manager.list_devices_by_category(Some("all")).await
+                        {
+                            let mut state = state_clone.lock().await;
+                            state.create_device_form.available_versions = targets.clone();
+                            state.create_device_form.available_device_types = devices.clone();
+
+                            {
+                                let mut cache = state.device_cache.write().await;
+                                cache.update_android_cache(devices, targets);
+                            }
+
+                            if let Some((id, display)) = state
+                                .create_device_form
+                                .available_device_types
+                                .first()
+                                .cloned()
+                            {
+                                state.create_device_form.device_type_id = id;
+                                state.create_device_form.device_type = display;
+                                state.create_device_form.selected_device_type_index = 0;
+                            }
+
+                            if let Some((value, display)) =
+                                state.create_device_form.available_versions.first().cloned()
+                            {
+                                state.create_device_form.version = value;
+                                state.create_device_form.version_display = display;
+                                state.create_device_form.selected_api_level_index = 0;
+                            }
+
+                            state.create_device_form.generate_placeholder_name();
+                            state.create_device_form.is_loading_cache = false;
+                        }
+                    }
+                }
+                Panel::Ios => {
+                    if let Some(ref ios_manager) = ios_manager {
+                        if let Ok(device_types) = ios_manager.list_device_types_with_names().await {
+                            if let Ok(runtimes) = ios_manager.list_runtimes().await {
+                                let mut state = state_clone.lock().await;
+                                state.create_device_form.available_device_types =
+                                    device_types.clone();
+                                state.create_device_form.available_versions = runtimes.clone();
+
+                                {
+                                    let mut cache = state.device_cache.write().await;
+                                    cache.update_ios_cache(device_types, runtimes);
+                                }
+
+                                if let Some((id, display)) = state
+                                    .create_device_form
+                                    .available_device_types
+                                    .first()
+                                    .cloned()
+                                {
+                                    state.create_device_form.device_type_id = id;
+                                    state.create_device_form.device_type = display;
+                                    state.create_device_form.selected_device_type_index = 0;
+                                }
+
+                                if let Some((value, display)) =
+                                    state.create_device_form.available_versions.first().cloned()
+                                {
+                                    state.create_device_form.version = value;
+                                    state.create_device_form.version_display = display;
+                                    state.create_device_form.selected_api_level_index = 0;
+                                }
+
+                                state.create_device_form.generate_placeholder_name();
+                                state.create_device_form.is_loading_cache = false;
+                            }
+                        }
+                    }
+                }
+            }
+        });
+    }
+
     #[allow(dead_code)]
     pub(super) async fn load_available_versions(&mut self) -> Result<()> {
         let state = self.state.lock().await;
@@ -112,6 +225,127 @@ impl App {
                         Some("iOS simulator not available on this platform.".to_string());
                     return Ok(());
                 }
+            }
+        }
+
+        Ok(())
+    }
+
+    pub(super) async fn handle_create_mode_key(&mut self, key: KeyEvent) -> anyhow::Result<()> {
+        match key.code {
+            KeyCode::Esc => {
+                let mut state = self.state.lock().await;
+                if !state.create_device_form.is_creating {
+                    state.mode = Mode::Normal;
+                    state.create_device_form.error_message = None;
+                }
+            }
+            KeyCode::Tab | KeyCode::Down => {
+                self.navigate_create_form(true).await;
+            }
+            KeyCode::BackTab | KeyCode::Up => {
+                self.navigate_create_form(false).await;
+            }
+            KeyCode::Enter => {
+                let is_creating = {
+                    let state = self.state.lock().await;
+                    state.create_device_form.is_creating
+                };
+                if !is_creating {
+                    self.submit_create_device().await?;
+                }
+            }
+            KeyCode::Left => {
+                self.change_create_device_selection(false).await?;
+            }
+            KeyCode::Right => {
+                self.change_create_device_selection(true).await?;
+            }
+            KeyCode::Char(c) if key.modifiers.contains(KeyModifiers::CONTROL) => {
+                self.handle_create_mode_ctrl_char(c).await?;
+            }
+            KeyCode::Char(c) => {
+                let mut state = self.state.lock().await;
+                if !state.create_device_form.is_creating {
+                    self.handle_create_device_char(&mut state, c);
+                }
+            }
+            KeyCode::Backspace => {
+                let mut state = self.state.lock().await;
+                if !state.create_device_form.is_creating {
+                    self.handle_create_device_backspace(&mut state);
+                }
+            }
+            _ => {}
+        }
+
+        Ok(())
+    }
+
+    async fn navigate_create_form(&mut self, forward: bool) {
+        let mut state = self.state.lock().await;
+        if state.create_device_form.is_creating {
+            return;
+        }
+
+        match (state.active_panel, forward) {
+            (Panel::Android, true) => state.create_device_form.next_field(),
+            (Panel::Android, false) => state.create_device_form.prev_field(),
+            (Panel::Ios, true) => state.create_device_form.next_field_ios(),
+            (Panel::Ios, false) => state.create_device_form.prev_field_ios(),
+        }
+    }
+
+    async fn change_create_device_selection(&mut self, move_right: bool) -> anyhow::Result<()> {
+        let should_reload = {
+            let mut state = self.state.lock().await;
+            if state.create_device_form.is_creating {
+                return Ok(());
+            }
+
+            let reload_category =
+                state.create_device_form.active_field == state::CreateDeviceField::Category;
+            let old_category = state.create_device_form.device_category_filter.clone();
+
+            if move_right {
+                self.handle_create_device_right(&mut state);
+            } else {
+                self.handle_create_device_left(&mut state);
+            }
+
+            reload_category && old_category != state.create_device_form.device_category_filter
+        };
+
+        if !should_reload {
+            return Ok(());
+        }
+
+        if let Err(error) = self.reload_device_types_for_category().await {
+            let mut state = self.state.lock().await;
+            state.create_device_form.error_message = Some(format_user_error(&error));
+        }
+
+        Ok(())
+    }
+
+    async fn handle_create_mode_ctrl_char(&mut self, c: char) -> anyhow::Result<()> {
+        let is_creating = {
+            let state = self.state.lock().await;
+            state.create_device_form.is_creating
+        };
+
+        if is_creating {
+            return Ok(());
+        }
+
+        match c {
+            'h' => self.change_create_device_selection(false).await?,
+            'l' => self.change_create_device_selection(true).await?,
+            'j' => self.navigate_create_form(true).await,
+            'k' => self.navigate_create_form(false).await,
+            _ => {
+                let mut state = self.state.lock().await;
+                self.handle_create_device_char(&mut state, c);
             }
         }
